Move PXIe_2597 to separate module, fix imports
